new high-res splash is working, but slow
