+++ src/components/Splash.vue
@@ -1,165 +1,172 @@
-  
 <template>
-  <section id="scrubbable">
-    <div id="container">
-      <div class="scrollDist mtn">
-        <div class="main mtn">
-          <svg
-            viewBox="0 0 1200 800"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <mask id="m">
-              <g class="cloud1">
-                <rect
-                  fill="#fff"
-                  width="100%"
-                  height="801"
-                  y="799"
-                />
-                <image
-                  xlink:href="@/assets/videos/cloudMask-01-01.png"
-                  width="1200"
-                  height="800"
-                />
-              </g>
-            </mask>
-                    
-            <!-- this grabbed from codepen and need to be replaced with our own images -->
-            <!-- <image class="mountBg" xlink:href="https://assets.codepen.io/721952/mountBg.png" width="100%" height="100%"/>   -->
-            <!-- <image class="mountBg" xlink:href="https://assets.codepen.io/721952/mountBg.png" width="100%" height="100%"/>   -->  
-            <image
-              class="mountMg"
-              xlink:href="@/assets/titleImages/splash/mountainMG.png"
-              width="100%"
-              height="100%"
-            />    
-            <image
-              class="cloud2"
-              xlink:href="https://assets.codepen.io/721952/cloud2.png"
-              width="100%"
-              height="100%"
-            />    
-            <image
-              class="mountFg"
-              xlink:href="@/assets/titleImages/splash/frozen-lakeFG.png"
-              width="100%"
-              height="100%"
-            />
-            <image
-              class="people"
-              xlink:href="@/assets/titleImages/splash/people.png"
-              width="100%"
-              height="100%"
-            />
-            <!-- <image class="mountFg" xlink:href="https://assets.codepen.io/721952/mountFg.png" width="100%" height="100%"/> -->
-            <image
-              class="cloud1"
-              xlink:href="https://assets.codepen.io/721952/cloud1.png"
-              width="100%"
-              height="100%"
-            />
-            <image
-              class="cloud3"
-              xlink:href="https://assets.codepen.io/721952/cloud3.png"
-              width="100%"
-              height="100%"
-            />
-            <text
-              class="overall-title"
-              fill="#fff"
-              x="600"
-              y="200"
-              text-anchor="middle"
-            >From Snow to Flow</text>
-            <text
-              class="byline"
-              x="280"
-              y="250"
-              fill="#fff"
-              text-anchor="left"
-            >U.S. Geological Survey<tspan
-              dx="-145"
-              dy="20"
-            >Water Resources Mission Area</tspan></text>
-
-                    
-            <g mask="url(#m)">
-              <rect
-                fill="#fff"
-                width="100%"
-                height="80%"
-              />      
-
-            </g>
-                    
-          </svg>
-        </div>
-      </div>
+  <div class="splash">
+    <div class="splashTitle">
+      <h1>From Snow to Flow</h1>
+      <p>U.S. Geological Survey</p>
+      <p>Water Resources Mission Area</p>
     </div>
-  </section>
+    <!-- <div class="splashOverlay" /> -->
+    <div
+      id="mountains"
+      class="element"
+      data-depth="0.10"
+    >
+      <picture>
+        <!-- <source srcset="@/assets/images/splash/1x/mountains-1x.png"> -->
+        <source srcset="@/assets/titleImages/splash/mountainMG.png">
+        <img 
+          src="@/assets/titleImages/splash/mountainMG.png"
+          xlink:href="@/assets/titleImages/splash/mountainMG.png"
+          >
+      </picture>
+    </div>
+    <div
+      id="water"
+      class="element"
+      data-depth="0.30"
+    >
+      <picture>
+        <!-- <source srcset="@/assets/images/splash/1x/water-1x.png"> -->
+        <source srcset="@/assets/titleImages/splash/frozen-lakeFG.png">
+        <img src="@/assets/titleImages/splash/frozen-lakeFG.png">
+      </picture>
+    </div>
+    <div
+      id="people"
+      class="element"
+      data-depth="0.20"
+    >
+      <picture>
+        <!-- <source srcset="@/assets/images/splash/1x/people-1x.png"> -->
+        <source srcset="@/assets/titleImages/splash/people.png">
+        <img src="@/assets/titleImages/splash/people.png">
+      </picture>
+    </div>
+    <div
+      id="clouds"
+      class="element cover"
+      data-depth="0.80"
+    >
+       <picture>
+         <source srcset="https://assets.codepen.io/721952/cloud2.png">
+         <img src="https://assets.codepen.io/721952/cloud2.png">
+       </picture>
+    </div>
+    <div
+      id="more-clouds"
+      class="element cover"
+      data-depth="0.90"
+    >
+       <picture>
+         <source srcset="@/assets/titleImages/splash/more-clouds.png">
+         <img src="@/assets/titleImages/splash/more-clouds.png">
+       </picture>
+    </div>
+  </div>
 </template>
-
 <script>
 import { ScrollToPlugin } from "gsap/ScrollToPlugin"; // to trigger scroll events
 import { ScrollTrigger } from "gsap/ScrollTrigger"; // animated scroll events
-    export default {
-        name: 'Splash',
-        data() {
-            return {
-                myVar: 50,
-            }
-        },
-        computed: {
-            windowHeight: function () {
-                const usgsBannerHeight = 85;
-                return Number(this.$store.state.windowHeight - this.$store.state.warningHeight - usgsBannerHeight) + 'px';
-            }
-        },
-        mounted() { 
-            const gsap = this.$gsap;
-            gsap.registerPlugin(ScrollToPlugin, ScrollTrigger); // register gsap plugins for scrollTrigger
-            // define starting posiitions of everything
-            gsap.set('.main', {position:'absolute', background:'#fff', width:'100%', maxWidth:'1800px', height:'70%', top:0, left:'50%', x:'-50%'})
-            gsap.set('.scrollDist', {width:'100%', height:'100%'})
-
-            // definite timeline of events
-            gsap.timeline({scrollTrigger:{trigger:'.scrollDist', start:'top top', end:'bottom bottom', scrub:1}})
-                // .fromTo('.sky', {y:-200},{y:0}, 0)
-                .fromTo('.cloud1', {y:100},{y:-800}, 0)
-                .fromTo('.cloud2', {y:200},{y:-500}, 0)
-                .fromTo('.cloud3', {y:30},{y:-650}, 0)
-                //.fromTo('.mountBg', {y:-10},{y:-100}, 0)
-                .fromTo('.mountMg', {y:-100},{y:-350}, 0)
-                .fromTo('.mountFg', {y:-50},{y:-600}, 0)
-/* 
-            $('#arrowBtn').on('mouseenter', (e)=>{ gsap.to('.arrow', {y:10, duration:0.8, ease:'back.inOut(3)', overwrite:'auto'}); })
-            $('#arrowBtn').on('mouseleave', (e)=>{ gsap.to('.arrow', {y:0, duration:0.5, ease:'power3.out', overwrite:'auto'}); })
-            $('#arrowBtn').on('click', (e)=>{ gsap.to(window, {scrollTo:innerHeight, duration:1.5, ease:'power1.inOut'}); }) // scrollTo requires the ScrollTo plugin (not to be confused w/ ScrollTrigger)
-       */  }
+export default {
+    name: "Splash",
+    mounted(){
+        this.$nextTick(() => this.splashParallax());
+    },
+    methods:{
+        splashParallax(){
+            const self = this;
+            this.$gsap.registerPlugin(ScrollToPlugin, ScrollTrigger); // register gsap plugins for scrollTrigger 
+                ScrollTrigger.matchMedia({
+                    //desktop 
+                    "(min-width: 800px)": function(){
+                        self.$gsap.timeline({
+                            scrollTrigger:{
+                                trigger: ".splash",
+                                start: "top top",
+                                end:`bottom center`,
+                                scrub: true
+                            }
+                        })
+                        .fromTo("#more-clouds", {yPercent: 40}, {yPercent: -20}, 0)
+                        .fromTo("#clouds", {yPercent: 40}, {yPercent: -10}, 0)
+                        .fromTo("#people", {yPercent: 19}, {yPercent: -15}, 0)
+                        .fromTo("#water", {yPercent: 15}, {yPercent: -15}, 0)
+                        .fromTo("#mountains", {yPercent: 10}, {yPercent: 0}, 0)
+                    }
+                })
+        }
     }
+}
 </script>
-
-<style scoped lang="scss">
-#scrubbable {
-  padding: 0 0 3em 0;
-}
-svg {
-   background-color: #0b5eb4;
-}
-.mtn {
-  position:relative;
- 
-}
-#container {
+<style lang="scss" scoped>
+.splash{
     position: relative;
-    height: 120vh;
-    width: 100%;
+    height: 80vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-image: linear-gradient(#0c5fb4 0%, #fff 50%);
 }
-#cheese {
-    position: fixed;
-    width: 80%;
-    top: 62%;
-    left: 25%;
-    margin:auto;
+.splashTitle{
+    position: absolute;
+    top: 10%;
+    z-index: 22;
+    margin-bottom: 20px;
+    padding: 0 15px;
+    h1{
+        font-size:clamp(5em, 10vw, 2em);
+        font-weight: 800;
+    }
+    h1, p{
+        color: #fff;
+    } 
+    p{
+        padding-top: 10px;
+        font-weight: bold;
+    }
+}
+.splashOverlay{
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background: rgba(0,0,0,.3);
+    z-index: 21
+}
+.element{
+    position: absolute;
+    width: 100%;
+    height: 100%;
+}
+.element img{
+    object-fit: contain;
+    object-position: center;
+    width: 100%;
+    height: 100%;
+}
+#mountains{
+    z-index: 10;
+}
+#water{
+    z-index: 15;
+}
+#people{
+    z-index: 20;
+}
+#clouds{
+    z-index: 25;
+    height: 125%;
+    // display: none;
+}
+#more-clouds {
+   z-index: 30;
+  height: 125%;
+}
+.cover {
+  object-fit: cover;
+}
+@media screen and (min-width:800px){
+    #clouds{
+        display: block;
+    }
 }
 </style>
